requirements.txt
diffusers
einops
matplotlib
numpy<2.*
pandas
Pillow
rich-click
setuptools
torch
torchvision
tqdm
transformers
